some stupid refactoring

+++ src/client/java/com/akicater/CUIConfig.java
@@ -5,6 +5,8 @@ import me.shedaniel.autoconfig.annotation.Config;
 import me.shedaniel.autoconfig.annotation.ConfigEntry;
 
 import static com.akicater.Cui.MODID;
+import static net.minecraft.text.Text.literal;
+import static net.minecraft.text.Text.translatable;
 
 @Config(name = MODID)
 public class CUIConfig implements ConfigData {
@@ -12,5 +14,6 @@ public class CUIConfig implements ConfigData {
     @ConfigEntry.ColorPicker
     public String color = "#00ff11";
     public boolean rainbow = false;
-    public float ranbowSpeed = 20;
+    public int ranbowSpeed = 20;
+
 }
